Data safety-oriented and correctness-oriented rewrite.
Several code paths were racy in the previous implementation.

While the code still does not handle exceptional conditions,
it is in much greater shape and much better tested now.
Interestinlyg enough, the codebase now partially parses
to-do items into objects that can later be queried by
attribute, which is the first step towards getting to
a UI enhancement to modify and filter tasks sensibly.

At this point the codebase is ready to receive a UI overhaul,
which comes next after finishing the to-do item object
parsing completely.